refactor(repo): BoardRepository 댓글 CRUD를 BoardRepositoryComments로 분할
836줄 → 601줄 (코딩스타일 800줄 상한 준수). 댓글 6개 메서드(getComments/
getCommentById/createComment/updateComment/setCommentStatus/setCommentPinned)
본문을 executeXxx 자유함수로 추출하고 클래스는 delegate만 유지.

- 공통 rethrowRepositoryError를 BoardRepositoryHelpers로 이관·공유
  (Application의 rethrowOrHandle 패턴과 정합)
- export 경로·메서드 시그니처 불변 → 호출부 변경 0
- 순수 이동, 기능 변경 없음
