+++ uniqn-mobile/src/repositories/supabase/BoardRepository.ts
@@ -14,7 +14,7 @@
 
 import { supabase } from '@/lib/supabase';
 import { logger } from '@/utils/logger';
-import { toError, isAppError } from '@/errors';
+import { toError } from '@/errors';
 import { handleSupabaseError, runRpc } from '@/utils/supabase';
 import type {
   BoardComment,
@@ -43,13 +43,20 @@ import type {
 import {
   TABLES,
   POST_COLUMNS,
-  COMMENT_COLUMNS,
   VOTE_COLUMNS,
   toBoardPost,
-  toBoardComment,
   togglePostVoteFallback,
   toggleCommentReactionFallback,
+  rethrowRepositoryError,
 } from './BoardRepositoryHelpers';
+import {
+  executeGetComments,
+  executeGetCommentById,
+  executeCreateComment,
+  executeUpdateComment,
+  executeSetCommentStatus,
+  executeSetCommentPinned,
+} from './BoardRepositoryComments';
 import {
   executeGetMembershipsByUser,
   executeGetMembershipsByPost,
@@ -66,25 +73,6 @@ import {
 // ============================================================================
 // Repository Implementation
 // ============================================================================
-
-/**
- * 캐치 블록 공통 처리: AppError는 그대로 전파하고, 그 외 오류는 로깅 후
- * handleSupabaseError로 위임한다. handleSupabaseError가 항상 throw하므로
- * 반환 타입은 never.
- */
-function rethrowRepositoryError(
-  error: unknown,
-  logMessage: string,
-  operation: string,
-  table: string,
-  context?: Record<string, unknown>
-): never {
-  if (isAppError(error)) throw error;
-  logger.error(logMessage, toError(error), context);
-  handleSupabaseError(error, { operation, table });
-  // handleSupabaseError는 항상 throw하지만 타입 시스템상 도달 불가 보장을 위해 재throw
-  throw error;
-}
 
 export class SupabaseBoardRepository implements IBoardRepository {
   // ==========================================================================
@@ -351,109 +339,15 @@ export class SupabaseBoardRepository implements IBoardRepository {
   // ==========================================================================
 
   async getComments(postId: string): Promise<BoardComment[]> {
-    try {
-      const { data, error } = await supabase
-        .from(TABLES.BOARD_COMMENTS)
-        .select(COMMENT_COLUMNS)
-        .eq('post_id', postId)
-        .order('created_at', { ascending: true });
-
-      if (error) {
-        handleSupabaseError(error, { operation: '댓글 조회', table: TABLES.BOARD_COMMENTS });
-      }
-
-      return ((data ?? []) as Record<string, unknown>[]).map(toBoardComment);
-    } catch (error) {
-      rethrowRepositoryError(error, '댓글 조회 실패', '댓글 조회', TABLES.BOARD_COMMENTS, {
-        postId,
-      });
-    }
+    return executeGetComments(postId);
   }
 
   async getCommentById(postId: string, commentId: string): Promise<BoardComment | null> {
-    try {
-      const { data, error } = await supabase
-        .from(TABLES.BOARD_COMMENTS)
-        .select(COMMENT_COLUMNS)
-        .eq('id', commentId)
-        .eq('post_id', postId)
-        .maybeSingle();
-
-      if (error) {
-        handleSupabaseError(error, { operation: '댓글 단건 조회', table: TABLES.BOARD_COMMENTS });
-      }
-
-      return data ? toBoardComment(data as Record<string, unknown>) : null;
-    } catch (error) {
-      rethrowRepositoryError(
-        error,
-        '댓글 단건 조회 실패',
-        '댓글 단건 조회',
-        TABLES.BOARD_COMMENTS,
-        {
-          postId,
-          commentId,
-        }
-      );
-    }
+    return executeGetCommentById(postId, commentId);
   }
 
   async createComment(input: CreateBoardCommentInput): Promise<string> {
-    try {
-      const now = new Date().toISOString();
-
-      // 댓글 생성
-      const { data, error } = await supabase
-        .from(TABLES.BOARD_COMMENTS)
-        .insert({
-          post_id: input.postId,
-          parent_comment_id: input.parentCommentId ?? null,
-          body: input.body,
-          author_id: input.authorId,
-          author_name: input.authorName,
-          author_role: input.authorRole,
-          mentioned_user_ids: input.mentionedUserIds ?? [],
-          reaction_counts: {},
-          is_pinned: false,
-          pinned_at: null,
-          pinned_by: null,
-          status: 'active',
-          image_attachments: input.imageAttachments ?? [],
-          created_at: now,
-          updated_at: now,
-        })
-        .select('id')
-        .single();
-
-      if (error) {
-        handleSupabaseError(error, { operation: '댓글 생성', table: TABLES.BOARD_COMMENTS });
-      }
-
-      // 게시글 댓글 수 + 활동 시각 업데이트
-      const { data: postData } = await supabase
-        .from(TABLES.BOARD_POSTS)
-        .select('comment_count')
-        .eq('id', input.postId)
-        .single();
-
-      if (postData) {
-        await supabase
-          .from(TABLES.BOARD_POSTS)
-          .update({
-            comment_count:
-              (((postData as Record<string, unknown>).comment_count as number) ?? 0) + 1,
-            last_activity_at: now,
-            updated_at: now,
-          })
-          .eq('id', input.postId);
-      }
-
-      return (data as Record<string, unknown>).id as string;
-    } catch (error) {
-      rethrowRepositoryError(error, '댓글 생성 실패', '댓글 생성', TABLES.BOARD_COMMENTS, {
-        postId: input.postId,
-      });
-    }
+    return executeCreateComment(input);
   }
 
   async updateComment(
@@ -461,29 +355,7 @@ export class SupabaseBoardRepository implements IBoardRepository {
     commentId: string,
     input: UpdateBoardCommentInput
   ): Promise<void> {
-    try {
-      const updates: Record<string, unknown> = {
-        updated_at: new Date().toISOString(),
-      };
-
-      if (input.body !== undefined) updates.body = input.body;
-      if (input.imageAttachments !== undefined) updates.image_attachments = input.imageAttachments;
-
-      const { error } = await supabase
-        .from(TABLES.BOARD_COMMENTS)
-        .update(updates)
-        .eq('id', commentId)
-        .eq('post_id', postId);
-
-      if (error) {
-        handleSupabaseError(error, { operation: '댓글 수정', table: TABLES.BOARD_COMMENTS });
-      }
-    } catch (error) {
-      rethrowRepositoryError(error, '댓글 수정 실패', '댓글 수정', TABLES.BOARD_COMMENTS, {
-        postId,
-        commentId,
-      });
-    }
+    return executeUpdateComment(postId, commentId, input);
   }
 
   async setCommentStatus(
@@ -491,86 +363,7 @@ export class SupabaseBoardRepository implements IBoardRepository {
     commentId: string,
     status: BoardComment['status']
   ): Promise<void> {
-    try {
-      const placeholder =
-        status === 'hidden' ? '관리자에 의해 숨김된 댓글입니다.' : '삭제된 댓글입니다.';
-      const now = new Date().toISOString();
-
-      // 댓글 현재 상태 확인
-      const { data: commentData, error: fetchError } = await supabase
-        .from(TABLES.BOARD_COMMENTS)
-        .select('status')
-        .eq('id', commentId)
-        .eq('post_id', postId)
-        .maybeSingle();
-
-      if (fetchError) {
-        handleSupabaseError(fetchError, {
-          operation: '댓글 상태 조회',
-          table: TABLES.BOARD_COMMENTS,
-        });
-      }
-
-      if (!commentData) return;
-
-      const previousStatus =
-        ((commentData as Record<string, unknown>).status as string) ?? 'active';
-      if (previousStatus !== 'active' || status === 'active') return;
-
-      // 댓글 상태 변경
-      const { error: updateError } = await supabase
-        .from(TABLES.BOARD_COMMENTS)
-        .update({
-          status,
-          body: placeholder,
-          is_pinned: false,
-          pinned_at: null,
-          pinned_by: null,
-          image_attachments: [],
-          mentioned_user_ids: [],
-          updated_at: now,
-        })
-        .eq('id', commentId)
-        .eq('post_id', postId);
-
-      if (updateError) {
-        handleSupabaseError(updateError, {
-          operation: '댓글 상태 변경',
-          table: TABLES.BOARD_COMMENTS,
-        });
-      }
-
-      // 게시글 댓글 수 감소
-      const { data: postData } = await supabase
-        .from(TABLES.BOARD_POSTS)
-        .select('comment_count')
-        .eq('id', postId)
-        .single();
-
-      if (postData) {
-        const currentCount = ((postData as Record<string, unknown>).comment_count as number) ?? 0;
-        await supabase
-          .from(TABLES.BOARD_POSTS)
-          .update({
-            comment_count: Math.max(0, currentCount - 1),
-            last_activity_at: now,
-            updated_at: now,
-          })
-          .eq('id', postId);
-      }
-    } catch (error) {
-      rethrowRepositoryError(
-        error,
-        '댓글 상태 변경 실패',
-        '댓글 상태 변경',
-        TABLES.BOARD_COMMENTS,
-        {
-          postId,
-          commentId,
-          status,
-        }
-      );
-    }
+    return executeSetCommentStatus(postId, commentId, status);
   }
 
   async setCommentPinned(
@@ -579,35 +372,7 @@ export class SupabaseBoardRepository implements IBoardRepository {
     isPinned: boolean,
     actorId: string
   ): Promise<void> {
-    try {
-      const now = new Date().toISOString();
-      const { error } = await supabase
-        .from(TABLES.BOARD_COMMENTS)
-        .update({
-          is_pinned: isPinned,
-          pinned_at: isPinned ? now : null,
-          pinned_by: isPinned ? actorId : null,
-          updated_at: now,
-        })
-        .eq('id', commentId)
-        .eq('post_id', postId);
-
-      if (error) {
-        handleSupabaseError(error, { operation: '댓글 고정 설정', table: TABLES.BOARD_COMMENTS });
-      }
-    } catch (error) {
-      rethrowRepositoryError(
-        error,
-        '댓글 고정 설정 실패',
-        '댓글 고정 설정',
-        TABLES.BOARD_COMMENTS,
-        {
-          postId,
-          commentId,
-          isPinned,
-        }
-      );
-    }
+    return executeSetCommentPinned(postId, commentId, isPinned, actorId);
   }
 
   // ==========================================================================
